requirements.txt
# Core dependencies for ERPDiff
# Note: torch is pre-installed in AutoDL base image -- do NOT pip install on AutoDL
# Run on AutoDL: pip install scipy pyyaml matplotlib -i https://pypi.tuna.tsinghua.edu.cn/simple

numpy==1.26.4
PyYAML==6.0.2
scipy==1.15.2
matplotlib==3.10.3
torch==2.7.0+cu128
